feat: refactor car property to device layer
