+++ app/src/main/java/com/thoughtworks/carapp/presentation/main/AutoHoldButton.kt
@@ -55,9 +55,9 @@ fun AutoHoldButton(
             Image(
                 painter = painterResource(
                     id = if (isOpened) {
-                        R.drawable.ic_auto_hold_indicator_opened
+                        R.drawable.ic_auto_hold_indicator_on
                     } else {
-                        R.drawable.ic_auto_hold_indicator_closed
+                        R.drawable.ic_auto_hold_indicator_off
                     }
                 ),
                 contentDescription = ""
