+++ app/src/main/res/drawable/ic_auto_hold_indicator_on.xml
